scripts/automated_ingestion/requirements.txt
boto3
pid
PyGithub
